+++ guibbon/image_viewer.py
@@ -3,7 +3,7 @@ import enum
 import math
 import tkinter as tk
 import types
-from typing import List, Any, Optional
+from typing import Any, Optional
 
 import cv2
 import numpy as np
@@ -57,10 +57,10 @@ class ImageViewer:
         self.canvas = tk.Canvas(master=self.frame, height=height, width=width, bg="gray10")
         self.canvas_shape_hw = (height, width)
 
-        self.imgtk = None
+        self.imgtk: ImageTk.PhotoImage
         self.onMouse: CallbackMouse = None
         self.modifier = ImageViewer.Modifier()
-        self.interactive_overlay_instance_list: List[Any] = []
+        self.interactive_overlay_instance_list: list[Any] = []
         self.mode: Optional[MODE] = None
 
         self.mouse_pan_calculator = mouse_pan.MousePan(ImageViewer.BUTTONNUM.RIGHT, on_drag=self.on_mouse_pan_drag, on_release=self.on_mouse_pan_release)
@@ -238,7 +238,7 @@ class ImageViewer:
 
     def set_img2can_matrix(self, img2can_matrix: TransformMatrix):
         self.img2can_matrix = img2can_matrix.copy()
-        self.can2img_matrix = np.linalg.inv(self.img2can_matrix)
+        self.can2img_matrix = np.linalg.inv(self.img2can_matrix).astype(float)
 
     def createInteractivePoint(
             self,