+++ scripts/deploy_docs.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-cd "$(dirname "$0")"
-
-source ../.venv/Scripts/activate
-python deploy_docs.py